Return a nil notifier if shutdown already started
It was tricky to detect cases where shutdown had started when you requested notifiers.

To help for that common case, the library now *returns a nil Notifier* if shutdown has already
reached the stage you are requesting a notifier for.

This is backwards compatible, but makes it much easier to test for such a case.